Use upgradeToBlockifiedProductGallery instead of replaceBlockWithProductGallery
