+++ plugins/woocommerce/client/blocks/assets/js/atomic/blocks/product-elements/product-image-gallery/upgrade-notice.tsx
@@ -3,29 +3,12 @@
  */
 import { __ } from '@wordpress/i18n';
 import { createInterpolateElement } from '@wordpress/element';
-import { select } from '@wordpress/data';
 import { UpgradeDowngradeNotice } from '@woocommerce/editor-components/upgrade-downgrade-notice';
-import { findBlock } from '@woocommerce/utils';
 
 /**
  * Internal dependencies
  */
-import metadata from './block.json';
-import { replaceBlockWithProductGallery } from '../../../../blocks/product-gallery/edit-utils';
-
-const upgradeToBlockifiedProductGallery = ( blockClientId: string ) => {
-	const blocks = select( 'core/block-editor' ).getBlocks();
-	const foundBlock = findBlock( {
-		blocks,
-		findCondition: ( block ) =>
-			block.name === metadata.name && block.clientId === blockClientId,
-	} );
-
-	if ( foundBlock ) {
-		return replaceBlockWithProductGallery( foundBlock.clientId );
-	}
-	return false;
-};
+import { upgradeToBlockifiedProductGallery } from '../../../../blocks/product-gallery/edit-utils';
 
 export const UpgradeNotice = ( {
 	blockClientId,
