+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/components/upgrade-product-image-gallery/index.tsx
@@ -11,7 +11,7 @@ import { findBlock } from '@woocommerce/utils';
 /**
  * Internal dependencies
  */
-import { replaceBlockWithProductGallery } from '../../../product-gallery/edit-utils';
+import { upgradeToBlockifiedProductGallery } from '../../../product-gallery/edit-utils';
 
 export const UpgradeProductImageGallery = () => {
 	const [ productImageGalleryBlock, setProductImageGalleryBlock ] =
@@ -49,10 +49,13 @@ export const UpgradeProductImageGallery = () => {
 			isDismissible={ false }
 			actionLabel={ buttonLabel }
 			onActionClick={ () => {
-				replaceBlockWithProductGallery(
-					productImageGalleryBlock.clientId
-				);
-				setProductImageGalleryBlock( null );
+				if (
+					upgradeToBlockifiedProductGallery(
+						productImageGalleryBlock.clientId
+					)
+				) {
+					setProductImageGalleryBlock( null );
+				}
 			} }
 			status="warning"
 		>
